+++ src/lib/iTin.Core.Hardware/iTin.Core.Hardware.Specification/iTin.Core.Hardware.Specification.Dmi/DMI/DmiClassCollection.cs
@@ -19,12 +19,10 @@ namespace iTin.Core.Hardware.Specification.Dmi
         /// <param name="parent">Estructura.</param>
         internal DmiClassCollection(DmiStructure parent) : base(new List<DmiClass>())
         {
-            SMBIOS smbios = DmiHelper.Smbios;
-            int smbiosVersion = smbios.Version;
-            SmbiosStructureCollection structures = smbios.Get((SmbiosStructure)(int)parent.Class);
+            SmbiosStructureCollection structures = parent.Context.Get((SmbiosStructure)(int)parent.Class);
             foreach (var structure in structures)
             {
-                Items.Add(new DmiClass(DmiStructureFactory.Create(structure, smbiosVersion)));
+                Items.Add(new DmiClass(DmiStructureFactory.Create(structure, parent.Context.Version)));
             }
         }
     }